Extract lib/zendesk_api/resource_ticket.rb
